# TINKERPOP-2349 Migrate JSON serialization to System.Text.Json in Gremlin.NET